More flexibility

+++ run.sh
@@ -6,7 +6,27 @@ cd $MY_DIR
 
 RUN_MODE=$1
 
-if [[ "$RUN_MODE" != "--skip-download" ]]; then
+DOWNLOAD=true
+GENERATE_MBTILES=true
+
+# Parse command line options
+while [[ "$#" -gt 0 ]]; do
+    case $1 in
+        --skip-download)
+            DOWNLOAD=false
+            ;;
+        --skip-mbtiles)
+            GENERATE_MBTILES=false
+            ;;
+        *)
+            echo "Unknown option: $1"
+            exit 1
+            ;;
+    esac
+    shift
+done
+
+if [ "$DOWNLOAD" = true ]; then
 	echo "Downloading..."
 	npm run download
 fi
@@ -20,29 +40,11 @@ fi
 echo "Converting to GeoJSON..."
 GEOCODING_SERVER_URL="https://photon.komoot.io/reverse" CLUSTERING_ARANGODB_URL=$CLUSTERING_ARANGODB_URL npm run prepare-geojson
 
-docker compose run --rm tippecanoe \
-  tippecanoe -Q -o /data/planet_lifts.mbtiles \
-		-f -z 15 -Z 5 --simplify-only-low-zooms \
-		--drop-densest-as-needed \
-		--named-layer=lifts:/data/mapboxgl_lifts.geojson;
+if [ -z "$CLUSTERING_ARANGODB_URL" ]; then
+	docker compose down
+fi
 
-docker compose run --rm tippecanoe \
-  tippecanoe -Q -o /data/planet_runs.mbtiles \
-		-f -z 15 -Z 9 --simplify-only-low-zooms \
-		--drop-densest-as-needed \
-	  --named-layer=runs:/data/mapboxgl_runs.geojson;
-
-docker compose run --rm tippecanoe \
-  tippecanoe -Q -o /data/ski_areas.mbtiles \
-		-f -z 15 -Z 0 -B 0 \
-		--named-layer=skiareas:/data/mapboxgl_ski_areas.geojson;
-
-docker compose run --rm tippecanoe \
-  tile-join -f --no-tile-size-limit -o /data/openskimap.mbtiles /data/ski_areas.mbtiles /data/planet_runs.mbtiles /data/planet_lifts.mbtiles;
-
-rm -Rf data/openskimap/*
-docker compose run --rm tippecanoe \
-  tile-join -e /data/openskimap/ /data/openskimap.mbtiles;
-
-docker compose down
-
+if [ "$GENERATE_MBTILES" = true ]; then
+	echo "Convert to MBTiles..."
+	./generate_mbtiles.sh
+fi
